+++ src/routes/index.ts
@@ -12,8 +12,7 @@ import logger from '../utils/logging';
 import { isDevelopment, RECAPTCHA_V3_SITE_KEY } from '../utils/process';
 import { decrypt, encrypt } from './encryption';
 import { optout } from './optout';
-//import { validate } from './recaptcha';
-import createAssessment from './recaptchav3'; 
+import createAssessment from './recaptcha'; 
 
 const router = express.Router();
 
@@ -92,7 +91,6 @@ const handleEmailPromptSubmission: RequestHandler<{}, z.infer<typeof EmailPrompt
     }
   }
 
-  //const success = await validate(recaptcha);
   const success = await createAssessment(recaptcha, 'email_prompt');
   if (!success) {
     res.render('index', {
@@ -178,12 +176,10 @@ router.get('/ops/healthcheck', (req, res, _next) => {
 });
 
 Handlebars.registerHelper('siteKeyInput', () => {
-  // return `<input type="hidden" name="recpatchaSiteKey" id="recpatchaSiteKey" value="${RECAPTCHA_SITE_KEY}">`;
   return `<input type="hidden" name="recpatchaSiteKey" id="recpatchaSiteKey" value="${RECAPTCHA_V3_SITE_KEY}">`;
 });
 
 Handlebars.registerHelper('recaptchaScript', () => {
-  //return `<script src="https://www.google.com/recaptcha/api.js?render=${RECAPTCHA_SITE_KEY}"></script>`;
   return `<script src="https://www.google.com/recaptcha/enterprise.js?render=${RECAPTCHA_V3_SITE_KEY}"></script>`;
 });
