+++ src/utils/process.ts
@@ -9,8 +9,6 @@ declare global {
       RECAPTCHA_SECRET: string;
       RECAPTCHA_PROJECT_ID: string;
       RECAPTCHA_V3_SITE_KEY: string;
-      RECAPTCHA_V3_API_KEY: object;
-      GOOGLE_APPLICATION_CREDENTIALS: string;
       GOOGLE_APPLICATION_CREDENTIALS_JSON: string;
 
       OPTOUT_ENDPOINT_URL: string;
@@ -47,7 +45,7 @@ export function normalizePort(val: any): string | number | boolean {
 export const PORT = (process.env.PORT && normalizePort(process.env.PORT)) || '3000';
 
 export const { 
-  RECAPTCHA_SITE_KEY, RECAPTCHA_SECRET, RECAPTCHA_PROJECT_ID, RECAPTCHA_V3_SITE_KEY, RECAPTCHA_V3_API_KEY, GOOGLE_APPLICATION_CREDENTIALS, GOOGLE_APPLICATION_CREDENTIALS_JSON, 
+  RECAPTCHA_SITE_KEY, RECAPTCHA_SECRET, RECAPTCHA_PROJECT_ID, RECAPTCHA_V3_SITE_KEY, GOOGLE_APPLICATION_CREDENTIALS_JSON, 
 } = process.env;
 
 export const { OPTOUT_API_KEY } = process.env;
